Fix float conversion

+++ SeraServer/src/main/java/com/cdtp/seraserver/contexts/ConfigContext.java
@@ -10,6 +10,6 @@ import lombok.Setter;
 @NoArgsConstructor
 @AllArgsConstructor
 public class ConfigContext {
-    private int period = -1;
-    private int timeout = -1 ;
+    private float period = -1;
+    private float timeout = -1 ;
 }
